+++ text-to-speech/src/test/java/com/ibm/watson/developer_cloud/text_to_speech/v1/CustomizationsIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 IBM Corp. All Rights Reserved.
+ * Copyright 2017 IBM Corp. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -73,15 +73,6 @@ public class CustomizationsIT extends WatsonServiceTest {
     model.setName(MODEL_NAME);
     model.setDescription(MODEL_DESCRIPTION);
     model.setLanguage(MODEL_LANGUAGE);
-
-    return model;
-  }
-
-  private CustomVoiceModel instantiateVoiceModelJapanese() {
-    final CustomVoiceModel model = new CustomVoiceModel();
-    model.setName(MODEL_NAME);
-    model.setDescription(MODEL_DESCRIPTION);
-    model.setLanguage(MODEL_LANGUAGE_JAPANESE);
 
     return model;
   }