+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Change Log
 ==========
 
+## Version 3.7.0
+
+_2017-03-19_
+
+ * New: Added Natural Language Understanding
+ * Fix: Text-to-Speech failed with combine of special characters, e.g Spanish + ";" bug text-to-speech #602
+ * Fix: Alchemy Language - CombinedResults missing Emotion #545
+
+
 ## Version 3.6.0
 
 _2017-03-07_